cabinet-search: remove usage of base-path
base-path is a configuration for the app-shell and the router and can't be used
in activities since the activity can also be loaded in a different context
with different paths.

The only working way to reference assets is relative to where the javascript is hosted
and not where the javascript is loaded from.

This should fix cabinet when loaded in nexus.