person: add a manual sync button
In case the user clicks the sync button we trigger a sync in the
backend and poll typesense until the sync timestamp is updated
or we fail after we hit a timeout (no UI error implemented yet).

This provides the entry point and token to the person view, so we
can talk to the APIs there. Also some refactoring for typesense
code so we can reuse the typesense service there as well.