Adjust for new group-less schema
Instead of having multiple study-person documents in typesense we now
again have only one document per person.

To support filtering on nested obejct properties we now use the new
"parent.{sub1: foo && sub2: bar}" syntax introduced with typesense v29.
We hook into the filter_by generation code and adjust the conditions for
the nested fields we want to filter in combination (studies only atm).

Due to the new syntax only working on exactly one level of nesting we have
to move the studies to the top level, and move the "status.text" to
"statusText" (other fields would need to be renamed too, but this is the only
one affected which we facet on).

While at it also move the applications to the top level, so things are
consistent.