typesense: use the search API for fetching a single document
With the backend getting more complex (splitting collections etc) I don't
want to implement everything in the proxy for the search APIs and all the rest
(different keys, different hanlding of merging results), so just use the
search API for everything.

This depends on that there is no caching enabled on the client side, but we have
disabled that anyway. Otherwise a manual sync would not update things.