Add complete dataset for nationality codes
Targetting passports for starters.
The codes are iso-3 with some extensions, see
https://en.wikipedia.org/wiki/Machine-readable_passport#Nationality_/_Citizenship_codes

We translate to iso-2 so we can use the builting js API for display names.
In addition hardcode display text for the extra codes and map legacy
codes to new ones.

One open question is if we want to merge the legacy codes into the new
ones, for example show "Germany (DEU)" even if the code is "D". It doesn't
do this for now as I'm not sure if that's wanted in all cases.

In the backend we'll do the same translation when filling typesense.