Add a custom input element for nationality codes
And replace the current enum element with it.

The goal of the element is for us to get the three character
code from the user, while the user can either copy it from
the document, or alternatively search by name.

This pulls in a toolkit update, so we can have multiple translations
with the LangMixin