Basic person data PDF export
Use jspdf-autotable to create a PDF of the person data.
The data mirrors what is shown in the person modal.

The multiple levels are modelled via two types of heading colors
with some indention on the left, to keep things simple.

Just a first draft to get a feeling for it.