Add a visual indicator for when the search takes a long time
Sometimes when the search takes a long time it's not clear if the
result list was already updated or if the search is just not done.

Use the typesense stalled indication to dim and disable the result
list if the search takes a long time.

We could also use other indicators, like a loading spinners etc.