facets: hide empty groups during rendering and not before
For hidden facets we still need to render them to the DOM which
means their groups are used and not filtered out. Move the group
removal logic into the rendering and hide the group if none of the
nested facets is visible instead.

This fixes the person group always showing due to the hidden
person.person facet.