Rework date refinement to handle timezones correctly
We have two types of datepickers, one for birthdate, where we shouldn't
handle timezones and just treat the input as UTC and not local time
(if the user enters 2024-12-24 the unix timestamp should also resolve
to the same day in UTC, which is used in the backend). The others are
normal and treat the input as local time and convert to unix timestamp.

To handle this we add a widgetParam to force UTC for the facet widget,
and introduce a new render function for the current refinements which
also does UTC.

To simplify things we use unix timestamps everywhere instead of JS timestamps,
so we only have to deal with two formats and not three.

We also have to make sure that converting to unix timestamps and back again
always result in the same date, to avoid getting in an update loop, breaking
keyboard input if a partial date input is set as refinement and then mirrored
back to the facet.

As a result the keyboard input of dates should now work, and birthdate
filtering should be more correct (though we use 12:00 for that in the backend,
so avoided timezones issues hackily that way).