date-range: debounce refine calls
This makes scrolling the year/month faster in firefox which emits
a change event on every scroll event.

Since this means we get a delayed callback from instantsearch about
the refinement change we need to ignore those updates while editing
is going on, to avoid resetting the date picker to an outdated value.