date-range: set the other date picker in case it wasn't set manually
In case a user enters a start or end date, prefill the other date as well,
to make entering single days more easy.

Since when typing a date the change event will fire multiple times during
entering a date keep track of the manuall entered state and keep the
dates in sync.

Since chaning the other date will now update the min/max of the currently
focused date, we have to ignore min/max updates when focused since that breaks
input in chrome (typing gets reset when min/max changes for some reason).
The min/max will be applied on blur or when the refinement callback comes back.