file hits: conditional highlighting of the additional type
We want to have highlighting of search results, but we only get that
for the language that is in typesense. To work around that do the
highlighting only if the value in typesense matches the frontend
translation of the current language. If not we just show the translation
and no highlighting.

Better than nothing...