Adjust for translation schema changes
Various fields in typesense now got english translations, so adjust for that.
Reload instantsearch/typesense on language changes to use the correct fields.

* Facet fields depend on the language
* Query by fields depend on the language, which means we can remove
  the highlight hack for the document types (renderFieldWithHighlightOrTranslated)
* Sorting depends on the language. There are now three sort fields in the backend
  to optimize memory a bit there.

With the fields they mostly just gained a "textEn" variants, and logic to
select that. The only field that changed more is file.base.studyField,
which got adjusted to match the structure of all other translated fields.
So now "file.base.studyField.key" maps to "studyField" in blob.