Add basic visual grouping of search hits
Remove the instantsearch provided styles and vendor some for
the search bar, and try to replicate them for the pagination etc.
This removes all the initial styling of the hits and gives us more
control there.

Move the border and spacing styling into the search hits, we can
control things there.

Pass to the hits if they are the first in a group so they can decide
wether to add spacing or not.

Hide some content from obsolete hits, so their height is smaller. Not
following the mockups yet, just to see how the layout looks like.