search: show a better error message in case there could be more results
Since we only can do something like 20 pages with the current backend, if the
user goes to a page after that we would show a "no results" message.
Instead explain that there could be more results, but we are not showing them.

In case the last page is not complete and there would be more results we don't
really know though. The total count from typesense is not exact, so we can't
know if there would be more.